refactor: download single and multiple champions
